fixed iterators
