+++ src/agdb/collections/vec.rs
@@ -1,2 +1,3 @@
 pub mod old_storage_vec_iterator;
-pub mod vec_storage_iterator;
+pub mod vec_storage_iterator_dynamic_sized;
+pub mod vec_storage_iterator_fixed_size;
